fix : rename NaverArticle to CafeArticle
