+++ src/main/java/com/sanhak/backend/domain/article/repository/CARepository.java
@@ -1,14 +1,14 @@
 package com.sanhak.backend.domain.article.repository;
 
-import com.sanhak.backend.domain.article.NaverArticle;
+import com.sanhak.backend.domain.article.CafeArticle;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface NARepository extends JpaRepository<NaverArticle, Long> {
-    Optional<NaverArticle> findById(Long id);
+public interface CARepository extends JpaRepository<CafeArticle, Long> {
+    Optional<CafeArticle> findById(Long id);
 
     void deleteById(Long id);
 }
