+++ src/main/java/com/sanhak/backend/domain/article/service/CAService.java
@@ -1,7 +1,7 @@
 package com.sanhak.backend.domain.article.service;
 
-import com.sanhak.backend.domain.article.NaverArticle;
-import com.sanhak.backend.domain.article.repository.NARepository;
+import com.sanhak.backend.domain.article.CafeArticle;
+import com.sanhak.backend.domain.article.repository.CARepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -9,16 +9,16 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class NAService {
-    private final NARepository naRepository;
+public class CAService {
+    private final CARepository CARepository;
 
-    public NaverArticle findById(Long id) {
-        return naRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+    public CafeArticle findById(Long id) {
+        return CARepository.findById(id).orElseThrow(IllegalArgumentException::new);
     }
 
     @Transactional
     public Long deleteById(Long id) {
-        naRepository.deleteById(id);
+        CARepository.deleteById(id);
         return id;
     }
 }
